컨트롤러 파일 하나로 병합
분리되어 있으니 개발 시 헷갈려서 하나로 합침
